+++ src/mapping_writer.h
@@ -415,7 +415,8 @@ void MappingWriter<MappingRecord>::ProcessAndOutputMappingsInLowMemory(
               std::min((uint32_t)std::numeric_limits<uint8_t>::max(),
                        num_last_mapping_dups);
           if (mapping_parameters_.Tn5_shift) {
-            last_mapping.Tn5Shift();
+            last_mapping.Tn5Shift(mapping_parameters_.Tn5_forward_shift,
+                                  mapping_parameters_.Tn5_reverse_shift);
           }
 
           AppendMapping(last_rid, reference, last_mapping);
@@ -464,7 +465,8 @@ void MappingWriter<MappingRecord>::ProcessAndOutputMappingsInLowMemory(
     last_mapping.num_dups_ = std::min(
         (uint32_t)std::numeric_limits<uint8_t>::max(), num_last_mapping_dups);
     if (mapping_parameters_.Tn5_shift) {
-      last_mapping.Tn5Shift();
+      last_mapping.Tn5Shift(mapping_parameters_.Tn5_forward_shift,
+                            mapping_parameters_.Tn5_reverse_shift);
     }
     AppendMapping(last_rid, reference, last_mapping);
     ++num_mappings_passing_filters;