+++ src/paf_mapping.h
@@ -46,11 +46,11 @@ class PAFMapping : public Mapping {
            std::tie(m.fragment_start_position_);
   }
   uint64_t GetBarcode() const { return 0; }
-  void Tn5Shift() {
+  void Tn5Shift(int forward_shift, int reverse_shift) {
     if (direction_ == 1) {
-      fragment_start_position_ += 4;
+      fragment_start_position_ += forward_shift;
     } else {
-      fragment_length_ -= 5;
+      fragment_length_ += reverse_shift;
     }
   }
   bool IsPositiveStrand() const { return direction_ > 0 ? true : false; }
@@ -179,11 +179,11 @@ class PairedPAFMapping : public Mapping {
            std::tie(m.fragment_start_position_, m.fragment_length_);
   }
   uint64_t GetBarcode() const { return 0; }
-  void Tn5Shift() {
-    fragment_start_position_ += 4;
-    positive_alignment_length_ -= 4;
-    fragment_length_ -= 9;
-    negative_alignment_length_ -= 5;
+  void Tn5Shift(int forward_shift, int reverse_shift) {
+    fragment_start_position_   += forward_shift;
+    positive_alignment_length_ -= forward_shift;
+    fragment_length_           -= (forward_shift - reverse_shift);
+    negative_alignment_length_ += reverse_shift;
   }
   bool IsPositiveStrand() const { return direction_ > 0 ? true : false; }
   uint32_t GetStartPosition() const {  // inclusive